Coneccion con orders
No pensaba subir esto aun porque no esta completo, pero decidi subirlo para no tener que preocuparme de que se borre.

Este codigo hace que al momento de presionar el boton de checkout la informacion en el carrito se suba a la tabla de orders en la DB.

Pero me falta anadir una columna nueva del arrival time y subir el producto.

Para manana ya deberia tener eso hecho y  solo tendria que hacer que el invoice se vea bien y despues la pagina de my orders

+++ frontend_controller/cartController.py
@@ -10,7 +10,7 @@ def addCartController(p_id, name, image, price, quantity, total, brand, category
     # Receive the variables that we got from POST originally and save in a dictItem to add to session cart
     # The add happens over at the cartModel
     dictitems = {p_id: {'name': name, 'image': image, 'price': price, 'quantity': int(quantity), 'brand': brand,
-                       'total_price': float(total), 'category': category, 'pound': pounds}}
+                       'total_price': int(total), 'category': category, 'pound': pounds}}
 
     return addCartModel(dictitems)
